HMRC-1714: Deploy without quantization for review
So the theory is that the CPU running in AWS might not be optimised for
the calculations that would happen in a quantized model and so
calculations on the model take a slower path.

Other than a pytorch upgrade this is the only change I can think of that
might negatively impact the inference speed.